@@ -15,65 +15,66 @@
   "dependencies": {
     "@turf/point-on-feature": "^6.5.0",
     "@turf/rhumb-destination": "^6.5.0",
-    "body-parser": "^1.19.0",
-    "class-transformer": "^0.2.3",
-    "class-transformer-validator": "^0.8.0",
-    "class-validator": "^0.11.0",
+    "body-parser": "^1.20.0",
+    "class-transformer": "^0.5.1",
+    "class-transformer-validator": "^0.9.1",
+    "class-validator": "^0.13.2",
     "compression": "^1.7.4",
-    "cookie-parser": "^1.4.4",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
-    "cross-env": "^7.0.0",
-    "envalid": "^6.0.1",
-    "express": "^4.17.1",
-    "faker": "^4.1.0",
-    "generate-password": "^1.6.0",
-    "helmet": "^3.21.2",
+    "cross-env": "^7.0.3",
+    "envalid": "^7.3.0",
+    "express": "^4.17.3",
+    "faker": "^6.6.6",
+    "generate-password": "^1.7.0",
+    "helmet": "^5.0.2",
     "jsonwebtoken": "^8.5.1",
-    "mkcert": "^1.4.0",
-    "morgan": "^1.9.1",
-    "nanoid": "^3.1.22",
-    "needle": "^2.6.0",
-    "nodemailer": "^6.5.0",
-    "passport": "^0.4.1",
+    "mkcert": "^1.5.0",
+    "morgan": "^1.10.0",
+    "nanoid": "^3.3.2",
+    "needle": "^3.0.0",
+    "nodemailer": "^6.7.3",
+    "passport": "^0.5.2",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "pg": "^8.3.2",
+    "pg": "^8.7.3",
     "pg-format": "^1.0.4",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "supertest": "^4.0.2",
-    "swagger-ui-express": "^4.1.6",
-    "uuid": "^7.0.2",
-    "winston": "^3.2.1"
+    "supertest": "^6.2.2",
+    "swagger-ui-express": "^4.3.0",
+    "uuid": "^8.3.2",
+    "winston": "^3.7.2"
   },
   "devDependencies": {
-    "@types/compression": "^1.7.0",
+    "@types/compression": "^1.7.2",
     "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.6",
-    "@types/express": "^4.17.2",
-    "@types/faker": "^4.1.9",
-    "@types/helmet": "0.0.45",
-    "@types/jest": "^25.1.1",
-    "@types/jsonwebtoken": "^8.3.7",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/faker": "^6.6.9",
+    "@types/helmet": "^4.0.0",
+    "@types/jest": "^27.4.1",
+    "@types/jsonwebtoken": "^8.5.8",
     "@types/mkcert": "^1.2.0",
-    "@types/morgan": "^1.7.37",
-    "@types/needle": "^2.5.1",
-    "@types/node": "^13.7.0",
-    "@types/nodemailer": "^6.4.1",
-    "@types/passport": "^1.0.2",
-    "@types/passport-http": "^0.3.8",
-    "@types/passport-jwt": "^3.0.3",
-    "@types/passport-local": "^1.0.33",
-    "@types/pg": "^7.14.9",
+    "@types/morgan": "^1.9.3",
+    "@types/needle": "^2.5.3",
+    "@types/node": "^17.0.23",
+    "@types/nodemailer": "^6.4.4",
+    "@types/passport": "^1.0.7",
+    "@types/passport-http": "^0.3.9",
+    "@types/passport-jwt": "^3.0.6",
+    "@types/passport-local": "^1.0.34",
+    "@types/pg": "^8.6.5",
     "@types/pg-format": "^1.0.2",
-    "@types/supertest": "^2.0.8",
-    "@types/swagger-ui-express": "^4.1.2",
-    "@types/uuid": "^7.0.0",
-    "prettier": "^1.19.1",
-    "ts-node": "^8.6.2",
-    "ts-node-dev": "^1.0.0-pre.44",
+    "@types/supertest": "^2.0.12",
+    "@types/swagger-ui-express": "^4.1.3",
+    "@types/uuid": "^8.3.4",
+    "@types/validator": "^13.7.2",
+    "prettier": "^2.6.2",
+    "ts-node": "^10.7.0",
+    "ts-node-dev": "^1.1.8",
     "tslint": "^6.1.3",
-    "typescript": "^3.7.5"
+    "typescript": "^4.6.3"
   },
   "resolutions": {
     "@turf/point-on-feature": "^6.5.0"
