@@ -1,5 +1,5 @@
-const fs = require('fs');
-const md5 = require('md5');
+import fs from 'fs';
+import md5 from 'md5';
 
 const insertQuotesInText = (key, value, carrReturn) => {
     let text = `${key}=${value}${carrReturn}`;
@@ -7,7 +7,7 @@ const insertQuotesInText = (key, value, carrReturn) => {
     return text;
 }
 
-module.exports = (osi4iotState) => {
+export default function(osi4iotState) {
 
     const secrets_dir = "./secrets"
     if (!fs.existsSync(secrets_dir)) fs.mkdirSync(secrets_dir);
@@ -21,8 +21,8 @@ module.exports = (osi4iotState) => {
         `ACCESS_TOKEN_LIFETIME=${osi4iotState.platformInfo.ACCESS_TOKEN_LIFETIME}\n`,
         `MQTT_SSL_CERTS_VALIDITY_DAYS=${osi4iotState.platformInfo.MQTT_SSL_CERTS_VALIDITY_DAYS}\n`,
         `ENCRYPTION_SECRET_KEY=${osi4iotState.platformInfo.ENCRYPTION_SECRET_KEY}\n`,
-        insertQuotesInText("PLATFORM_ADMIN_FIRST_NAME", osi4iotState.platformInfo.PLATFORM_ADMIN_FIRST_NAME,"\n"),
-        insertQuotesInText("PLATFORM_ADMIN_SURNAME", osi4iotState.platformInfo.PLATFORM_ADMIN_SURNAME,"\n"),
+        insertQuotesInText("PLATFORM_ADMIN_FIRST_NAME", osi4iotState.platformInfo.PLATFORM_ADMIN_FIRST_NAME, "\n"),
+        insertQuotesInText("PLATFORM_ADMIN_SURNAME", osi4iotState.platformInfo.PLATFORM_ADMIN_SURNAME, "\n"),
         `PLATFORM_ADMIN_USER_NAME=${osi4iotState.platformInfo.PLATFORM_ADMIN_USER_NAME}\n`,
         `PLATFORM_ADMIN_EMAIL=${osi4iotState.platformInfo.PLATFORM_ADMIN_EMAIL}\n`,
         `PLATFORM_ADMIN_PASSWORD=${osi4iotState.platformInfo.PLATFORM_ADMIN_PASSWORD}\n`,
@@ -34,18 +34,18 @@ module.exports = (osi4iotState) => {
         `NOTIFICATIONS_EMAIL_PASSWORD=${osi4iotState.platformInfo.NOTIFICATIONS_EMAIL_PASSWORD}\n`,
         `MAIN_ORGANIZATION_TELEGRAM_CHAT_ID=${osi4iotState.platformInfo.MAIN_ORGANIZATION_TELEGRAM_CHAT_ID}\n`,
         `MAIN_ORGANIZATION_TELEGRAM_INVITATION_LINK=${osi4iotState.platformInfo.MAIN_ORGANIZATION_TELEGRAM_INVITATION_LINK}\n`,
-        `TELEGRAM_BOTTOKEN=${osi4iotState.platformInfo.TELEGRAM_BOTTOKEN}\n`
+        `TELEGRAM_BOTTOKEN=${osi4iotState.platformInfo.TELEGRAM_BOTTOKEN}\n`,
     ];
 
     if (fs.existsSync('./secrets/admin_api.txt')) {
         fs.rmSync('./secrets/admin_api.txt');
     }
 
-    for (iorg = 1; iorg <= osi4iotState.certs.mqtt_certs.organizations.length; iorg++) {
+    for (let iorg = 1; iorg <= osi4iotState.certs.mqtt_certs.organizations.length; iorg++) {
         const num_master_devices = osi4iotState.certs.mqtt_certs.organizations[iorg - 1].master_devices.length;
         adminApiSecrets.push(`ORG_${iorg}_HASH=${osi4iotState.certs.mqtt_certs.organizations[iorg - 1].org_hash}\n`)
         let orgMasterDevicesHashes = `ORG_${iorg}_MASTER_DEVICE_HASHES=`;
-        for (idev = 1; idev <= num_master_devices; idev++) {
+        for (let idev = 1; idev <= num_master_devices; idev++) {
             const mdHash = osi4iotState.certs.mqtt_certs.organizations[iorg - 1].master_devices[idev - 1].md_hash;
             orgMasterDevicesHashes = orgMasterDevicesHashes + mdHash;
             if (idev < num_master_devices) orgMasterDevicesHashes = orgMasterDevicesHashes + ",";
